Add cx threading support for mbedtls
mbedTLS 3.1 made the PSA crypto library non-optional. However, PSA
uses a global key store that is not thread safe unless
MBEDTLS_THREADING_C is defined. This violates the convention in earlier
versions that the library is safe to use in multiple threads so long
as they do not share any mbedtls contexts.

mbedTLS only provides threading support for pthreads, which of course
does not exist on all our supported platforms.

This commit implements the needed functions for thread safety using
CX cross-platform mutexes.