Fix mbedTLS mutex hardlocks in debug builds
DEBUG_LEVEL definition needs to be passed down to mbedTLS now so that
it knows the proper size for the Mutex structure. This is slightly
tricky because cx depends on mbedtls rather than the other way around.

Possible TODO: Move these and some other important definitions into a
pseudo-target that both can depend on.