Fix edge case when appending to ropes
String concatenation attempts to reuse the destination buffer rather
than reallocate. However this fails if the destination is also one of
the source strings.

In most common case -- i.e. appending to a string -- this is already
handled by the optimization for that case. However, in some cases where
the optimization cannot be used, this ends up overwriting one of the
source strings too early.

The main trigger for this was appending to small string to a rope,
with the resulting total length being less than the rope threshold.
This is unusual but can happen if a rope was created as the result of
a substring operation -- pathSetExt being the culprit in the case where
this bug was identified.

Detect this case and allocate a temporary destiation when it's not
safe to reuse it. This also fixes the previously unsupported use case of
using strNConcat to prepend strings.