Add strPC
This function behaves like the old strC used to. It requires a pointer
to a string (NOT a strref!). In exchange, it returns a C string that
is guarateed to persist as long as the underlying string object remains
untouched.

This is unlike strC which can return a scratch buffer (if the string
is a rope), or strBuffer which ALWAYS duplicates strings that have
more than a single reference.